requirements.txt
imagehash
Pillow
numpy
opencv-python-headless
httpx
flask
pytest
